refactor : MainPage => MainSection으로 파일명 변경 및 커스텀훅 사용
